+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Added support for resolving 'ap' URIs with `gateways` parameter to `ap fetch` command.
+- Added support for resolving 'ap' URIs with `gateways` parameter to `ap fetch` and `ap import` commands.
+- Added support for resolving 'ap' URIs with `gateways` parameter to `/api/v2/search` API endpoint.
 
 ## [5.7.1] - 2026-07-12